pyproject.toml
[project]
name = "devsoftwaresolutionprototypes"
version = "0.1.0"
description = ""
authors = [
    {name = "Yagunov Denis", email = "spb.jagunov.denis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry.dependencies]
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.39"
aiosqlite = "^0.21.0"
pyjwt = "^2.10.1"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.1"
sqlalchemy-stubs = "^0.4"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
